+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Verb Changelog
 ## **main** (MELPA)
+- Added support for `Verb-Proxy` heading property.
 
 ## **2.16.0** - 2024-03-02 (MELPA Stable)
 - Fixed LF being used instead of CRLF in multipart boundaries.